Add base for the home screen
