+++ manager/src/pages/Home/styled.ts
@@ -16,16 +16,18 @@
 
 import styled from 'styled-components';
 
-const Content = styled.section`
-  width: 100%;
-  height: calc(100vh - 36px);
-  position: relative;
+const Wrapper = styled.section`
   display: flex;
-  justify-content: center;
-  align-items: center;
   flex-direction: column;
+  width: 100%;
+`;
+
+const Content = styled.div`
+  height: calc(100vh - 36px);
+  width: 100%;
 `;
 
 export default {
   Content,
+  Wrapper,
 };
