+++ manager/src/routes/index.tsx
@@ -17,6 +17,7 @@
 import React, { Suspense, lazy } from 'react';
 import { BrowserRouter, Redirect, Route, Switch } from 'react-router-dom';
 import { isLogged } from 'helpers/localStorage/tokens';
+import { PrivateRoute } from 'components';
 
 const MANAGER_PATH = (window as any).REACT_APP_HORUSEC_MANAGER_PATH || '/';
 
@@ -30,15 +31,22 @@ const Routes = () => (
 
         <Route path="/auth" component={lazy(() => import('pages/Auth'))} />
 
-        <Route
-          exact
+        <PrivateRoute
+          exact={true}
           path="/home"
           component={lazy(() => import('pages/Home'))}
         />
 
-        <Route
+        <PrivateRoute
+          exact={false}
           path="/overview"
           component={lazy(() => import('pages/Overview'))}
+        />
+
+        <PrivateRoute
+          path="/settings"
+          exact={true}
+          component={lazy(() => import('pages/Settings'))}
         />
 
         <Route component={lazy(() => import('pages/NotFound'))} />
